decode_to_md5.c: Silence warning with cast.
More size_t/unsigned int disagreement.

+++ examples/decode_to_md5.c
@@ -115,7 +115,7 @@ int main(int argc, char **argv) {
     size_t frame_size = 0;
     const unsigned char *frame = vpx_video_reader_get_frame(reader,
                                                             &frame_size);
-    if (vpx_codec_decode(&codec, frame, frame_size, NULL, 0))
+    if (vpx_codec_decode(&codec, frame, (unsigned int)frame_size, NULL, 0))
       die_codec(&codec, "Failed to decode frame");
 
     while ((img = vpx_codec_get_frame(&codec, &iter)) != NULL) {
